This is a Pac-Man game made in C via the SDL2 library.
Here, the motion of each ghost is completely random. They won't chase you.

To run the game, double click on "Pac-Man.exe" executable file.

As the game window appears, maximize it then press button "R" on the keyboard to refresh the window.

From there, select any game option displayed by pressing on the number next to it. E.g Press button "1" to play or press button "2" to quit and so on.

Press "P" during the game to pause it.
Press "N" during the game to change the background music.
Press "M" to go back to main menu.

PS: I do not own any of the songs or sfx available in the game.

For sure,  this game can still be ameliorated. Hope you'll have fun.